+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 ### Added
+* NetCDF plugin now support the size_of attribute, fixes
+  [#446](https://gitlab.maisondelasimulation.fr/pdidev/pdi/-/issues/446)
 
 ### Changed
 * Replaced the astyle based indentation by a clang-format one, fixes